+++ UnitTests/TestSuites/ColorsTests.swift
@@ -89,12 +89,24 @@ class ColorsTextFileTests: XCTestCase {
     XCTDiffStrings(result, expected)
   }
 
-  func testFileWithBadFormatting() {
+  func testFileWithBadSyntax() {
     let parser = ColorsTextFileParser()
     do {
-      try parser.parseFile(at: fixture("colors-bad.txt"))
-      XCTFail("Code did parse file successfully while it was expected to fail for bad formatting")
-    } catch ColorsParserError.invalidHexColor(string: ":", key: "MX_WELCOME_BACKGROUND"?) {
+      try parser.parseFile(at: fixture("colors-bad-syntax.txt"))
+      XCTFail("Code did parse file successfully while it was expected to fail for bad syntax")
+    } catch ColorsParserError.invalidFile {
+      // That's the expected exception we want to happen
+    } catch let error {
+      XCTFail("Unexpected error occured while parsing: \(error)")
+    }
+  }
+
+  func testFileWithBadValue() {
+    let parser = ColorsTextFileParser()
+    do {
+      try parser.parseFile(at: fixture("colors-bad-value.txt"))
+      XCTFail("Code did parse file successfully while it was expected to fail for bad value")
+    } catch ColorsParserError.invalidHexColor(string: "thisIsn'tAColor", key: "ArticleTitle"?) {
       // That's the expected exception we want to happen
     } catch let error {
       XCTFail("Unexpected error occured while parsing: \(error)")
@@ -119,7 +131,8 @@ class ColorsCLRFileTests: XCTestCase {
 
   func testFileWithDefaults() {
     let parser = ColorsCLRFileParser()
-    parser.parseFile(at: fixture("colors.clr"))
+    try! parser.parseFile(at: fixture("colors.clr"))
+
 
     let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
     let result = try! template.render(parser.stencilContext())
@@ -130,13 +143,25 @@ class ColorsCLRFileTests: XCTestCase {
 
   func testFileWithCustomName() {
     let parser = ColorsCLRFileParser()
-    parser.parseFile(at: fixture("colors.clr"))
+    try! parser.parseFile(at: fixture("colors.clr"))
 
     let template = GenumTemplate(templateString: fixtureString("colors-default.stencil"))
     let result = try! template.render(parser.stencilContext(enumName: "XCTColors"))
 
     let expected = fixtureString("Colors-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
+  }
+
+  func testFileWithBadFile() {
+    let parser = ColorsCLRFileParser()
+    do {
+      try parser.parseFile(at: fixture("colors-bad.clr"))
+      XCTFail("Code did parse file successfully while it was expected to fail for bad file")
+    } catch ColorsParserError.invalidFile {
+      // That's the expected exception we want to happen
+    } catch let error {
+      XCTFail("Unexpected error occured while parsing: \(error)")
+    }
   }
 }
 
@@ -182,7 +207,33 @@ class ColorsXMLFileTests: XCTestCase {
     let expected = fixtureString("Colors-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
   }
+
+  func testFileWithBadSyntax() {
+    let parser = ColorsXMLFileParser()
+    do {
+      try parser.parseFile(at: fixture("colors-bad-syntax.xml"))
+      XCTFail("Code did parse file successfully while it was expected to fail for bad syntax")
+    } catch ColorsParserError.invalidFile {
+      // That's the expected exception we want to happen
+    } catch let error {
+      XCTFail("Unexpected error occured while parsing: \(error)")
+    }
+  }
+
+  func testFileWithBadValue() {
+    let parser = ColorsXMLFileParser()
+    do {
+      try parser.parseFile(at: fixture("colors-bad-value.xml"))
+      XCTFail("Code did parse file successfully while it was expected to fail for bad value")
+    } catch ColorsParserError.invalidHexColor(string: "this isn't a color", key: "ArticleTitle"?) {
+      // That's the expected exception we want to happen
+    } catch let error {
+      XCTFail("Unexpected error occured while parsing: \(error)")
+    }
+  }
 }
+
+// MARK: - Tests for JSON color files
 
 class ColorsJSONFileTests: XCTestCase {
   func testEmpty() {
@@ -223,5 +274,29 @@ class ColorsJSONFileTests: XCTestCase {
 
     let expected = fixtureString("Colors-File-CustomName.swift.out")
     XCTDiffStrings(result, expected)
+  }
+
+  func testFileWithBadSyntax() {
+    let parser = ColorsJSONFileParser()
+    do {
+      try parser.parseFile(at: fixture("colors-bad-syntax.json"))
+      XCTFail("Code did parse file successfully while it was expected to fail for bad syntax")
+    } catch ColorsParserError.invalidFile {
+      // That's the expected exception we want to happen
+    } catch let error {
+      XCTFail("Unexpected error occured while parsing: \(error)")
+    }
+  }
+
+  func testFileWithBadValue() {
+    let parser = ColorsJSONFileParser()
+    do {
+      try parser.parseFile(at: fixture("colors-bad-value.json"))
+      XCTFail("Code did parse file successfully while it was expected to fail for bad value")
+    } catch ColorsParserError.invalidHexColor(string: "this isn't a color", key: "ArticleTitle"?) {
+      // That's the expected exception we want to happen
+    } catch let error {
+      XCTFail("Unexpected error occured while parsing: \(error)")
+    }
   }
 }